requirements.txt
ansible
PyYaml
